neurobridge/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
